requirements_cpu.txt
Flask
Flask-Cors
pyyaml
loguru
numpy==1.23.5
nltk
Unidecode
inflect
librosa==0.9.2
torch==2.2.2 --extra-index-url https://download.pytorch.org/whl/cpu --pre
torchaudio==2.2.2 --extra-index-url https://download.pytorch.org/whl/cpu --pre
soundfile==0.10.3.post1
flaskwebgui==1.1.6
waitress
briefcase
opencv-python
Pillow
scikit-image
kornia
yacs
torchvision==0.17.2 --extra-index-url https://download.pytorch.org/whl/cpu --pre
pydub
gfpgan
cog
imageio==2.19.3
imageio-ffmpeg==0.4.7
av
torch-optimizer
realesrgan
onnxruntime
insightface
pypinyin
einops
praat-parselmouth
denoiser
noisereduce
openunmix
torchlibrosa
apscheduler
httpx
cryptography
anyio==3.7.1
faster-whisper
whisper-timestamped
gTTS
scenedetect[opencv]
pykalman==0.9.7
decord
flask-babel
flasgger
peft
packaging
skbase